make locale handling more flexible and consistent with the other enum APIs.
we no longer explicitly enumerate pairs of locale / region. note that the hi-HI locale
has been eliminated as a side-effect; HI is not a region code that I could find, and
it is not mentioned in the Alexa docs. if we actually encounter hi-HI as a value, it
will decode to Language::Indian, Region::Other("HI"), which seems appropriate.

to support the locale refactor, i made the enum macro more support explicit string mappings.
this allowed me to use it in more places (e.g. IntentType).

I've added a few test cases to cover the new code, but will likely add more after looking
at the coverage report.